reimplemented how uploaded files are handled by multer, fixed deleting files and folders

+++ server/src/controllers/file/deleteFile.ts
@@ -5,6 +5,7 @@ import { AppError } from '../../errors/AppError.ts';
 import { parse as vParse } from 'valibot';
 import { ObjectIdSchema } from '../../validation-schemas/validationSchemas.ts';
 import User from '../../db/models/User.ts';
+import { deleteFolderRec } from '../../utils/deleteFolderRec.ts';
 
 export const deleteFile = async (
   req: Request,
@@ -18,25 +19,51 @@ export const deleteFile = async (
     // check if id is in correct format
     const fileId = vParse(ObjectIdSchema, req.params.id);
 
-    //delete from db
-    const file = await File.findOneAndDelete({
+    //check if file exists
+    const file = await File.findOne({
       userId: req.user.id,
       _id: fileId,
     }).exec();
 
     if (!file) {
-      throw new AppError('Could not delete file', 400);
+      throw new AppError('File not found', 404);
     }
 
-    //delete from disk
-    await Deno.remove(file.path);
+    // check if it is file or folder
+    if (file.fileType === 'folder') {
+      // delete from db recurisvely
+      const storageFreed = await deleteFolderRec(file.id);
 
-    // update used storage
-    await User.findByIdAndUpdate(req.user.id, {
-      usedStorage: req.user.usedStorage - file.size,
-    });
+      // update used storage
+      await User.findByIdAndUpdate(req.user.id, {
+        usedStorage: req.user.usedStorage - storageFreed,
+      });
 
-    res.status(200).send('Deleted file of id:' + file.id);
+      // delete from disk recursively
+      await Deno.remove(file.path, { recursive: true });
+
+      return void res.status(200).send('Deleted folder of id:' + file.id);
+    } else if (file.fileType === 'file') {
+      // delete from db
+      const deletedFile = await File.findOneAndDelete({
+        userId: req.user.id,
+        _id: fileId,
+      }).exec();
+
+      if (!deletedFile) {
+        throw new AppError('Could not delete file', 400);
+      }
+
+      // delete from disk
+      await Deno.remove(file.path);
+
+      // update used storage
+      await User.findByIdAndUpdate(req.user.id, {
+        usedStorage: req.user.usedStorage - file.size,
+      });
+
+      return void res.status(200).send('Deleted file of id:' + file.id);
+    }
   } catch (err) {
     next(err);
   }
